feat(types): export more types
- fix `getApksignerForOs` return type
- move options interface into its own file
- export `ApkSigningCommands` interface and attach to `ADB`
- export everything in `apks-utils` module; fix return values
- fix types in `tools/apk-signing`
- use TS v5.0.x
