+++ lib/adb.ts
@@ -3,8 +3,7 @@ import os from 'node:os';
 import methods, {getAndroidBinaryPath} from './tools';
 import {DEFAULT_ADB_EXEC_TIMEOUT, requireSdkRoot, getSdkRootFromEnv} from './helpers';
 import log from './logger';
-import {StringRecord} from '@appium/types';
-import type Logcat from './logcat';
+import type {ADBOptions, ADBExecutable} from './options';
 
 const DEFAULT_ADB_PORT = 5037;
 export const DEFAULT_OPTS = {
@@ -19,42 +18,6 @@ export const DEFAULT_OPTS = {
   allowOfflineDevices: false,
   allowDelayAdb: true,
 } as const;
-
-export interface ADBOptions {
-  sdkRoot?: string;
-  udid?: string;
-  appDeviceReadyTimeout?: number;
-  useKeystore?: boolean;
-  keystorePath?: string;
-  keystorePassword?: string;
-  keyAlias?: string;
-  keyPassword?: string;
-  executable: ADBExecutable;
-  tmpDir?: string;
-  curDeviceId?: string;
-  emulatorPort?: number;
-  logcat?: Logcat;
-  binaries?: StringRecord;
-  instrumentProc?: string;
-  suppressKillServer?: boolean;
-  jars?: StringRecord;
-  adbPort?: number;
-  adbHost?: string;
-  adbExecTimeout?: number;
-  remoteAppsCacheLimit?: number;
-  buildToolsVersion?: string;
-  allowOfflineDevices?: boolean;
-  allowDelayAdb?: boolean;
-  remoteAdbHost?: string;
-  remoteAdbPort?: number;
-  clearDeviceLogsOnStart?: boolean;
-  emPort?: number;
-}
-
-export interface ADBExecutable {
-  path: string;
-  defaultArgs: string[];
-}
 
 export class ADB {
   adbHost?: string;
