+++ lib/helpers.js
@@ -246,7 +246,7 @@ const getOpenSslForOs = async function () {
  * Get the absolute path to apksigner tool
  *
  * @param {Object} sysHelpers - An instance containing systemCallMethods helper methods
- * @returns {string} An absolute path to apksigner tool.
+ * @returns {Promise<string>} An absolute path to apksigner tool.
  * @throws {Error} If the tool is not present on the local file system.
  */
 async function getApksignerForOs (sysHelpers) {
