+++ lib/mixins.ts
@@ -5,24 +5,32 @@
  * @module
  */
 
-import {SystemCalls} from './tools/system-calls';
-import {ApkUtils} from './tools/apk-utils';
-import {ADBCommands} from './tools/adb-commands';
-import {SettingsClientCommands} from './tools/settings-client-commands';
-import {ADBEmuCommands} from './tools/adb-emu-commands';
-import {LockManagementCommands} from './tools/lockmgmt';
-import {ManifestMethods} from './tools/android-manifest';
-import {KeyboardCommands} from './tools/keyboard-commands';
+import {
+  SystemCalls,
+  ApkUtils,
+  ADBCommands,
+  SettingsClientCommands,
+  ADBEmuCommands,
+  LockManagementCommands,
+  ManifestMethods,
+  KeyboardCommands,
+  ApkSigningCommands,
+  ApksUtils,
+} from './tools';
+import {ADBOptions} from './options';
+
 declare module './adb' {
   // note that ADBOptions is the options object, but it's mixed directly in to the instance in the constructor.
   interface ADB
     extends ADBCommands,
       ApkUtils,
+      ApksUtils,
       SystemCalls,
       ADBOptions,
       SettingsClientCommands,
       ADBEmuCommands,
       LockManagementCommands,
       ManifestMethods,
-      KeyboardCommands {}
+      KeyboardCommands,
+      ApkSigningCommands {}
 }
