+++ lib/tools/apk-signing.js
@@ -31,8 +31,8 @@ const apkSigningMethods = {};
 /**
  * Execute apksigner utility with given arguments.
  *
- * @param {?Array<String>} args - The list of tool arguments.
- * @return {string} - Command stdout
+ * @param {string[]} args - The list of tool arguments.
+ * @return {Promise<string>} - Command stdout
  * @throws {Error} If apksigner binary is not present on the local file system
  *                 or the return code is not equal to zero.
  */
@@ -176,7 +176,7 @@ apkSigningMethods.sign = async function sign (appPath) {
  * Perform zip-aligning to the given local apk file.
  *
  * @param {string} apk - The full path to the local apk file.
- * @returns {boolean} True if the apk has been successfully aligned
+ * @returns {Promise<boolean>} True if the apk has been successfully aligned
  * or false if the apk has been already aligned.
  * @throws {Error} If zip-align fails.
  */
@@ -212,7 +212,7 @@ apkSigningMethods.zipAlignApk = async function zipAlignApk (apk) {
 
 /**
  * @typedef {Object} CertCheckOptions
- * @property {boolean} requireDefaultCert [true] Whether to require that the destination APK
+ * @property {boolean} [requireDefaultCert=true] Whether to require that the destination APK
  * is signed with the default Appium certificate or any valid certificate. This option
  * only has effect if `useKeystore` property is unset.
  */
@@ -223,7 +223,7 @@ apkSigningMethods.zipAlignApk = async function zipAlignApk (apk) {
  * @param {string} appPath - The full path to the local .apk(s) file.
  * @param {string} pgk - The name of application package.
  * @param {CertCheckOptions} opts - Certificate checking options
- * @return {boolean} True if given application is already signed.
+ * @return {Promise<boolean>} True if given application is already signed.
  */
 apkSigningMethods.checkApkCert = async function checkApkCert (appPath, pkg, opts = {}) {
   log.debug(`Checking app cert for ${appPath}`);
@@ -309,16 +309,16 @@ apkSigningMethods.checkApkCert = async function checkApkCert (appPath, pkg, opts
 
 /**
  * @typedef {Object} KeystoreHash
- * @property {?string} md5 the md5 hash value of the keystore
- * @property {?string} sha1 the sha1 hash value of the keystore
- * @property {?string} sha256 the sha256 hash value of the keystore
- * @property {?string} sha512 the sha512 hash value of the keystore
+ * @property {string} [md5] the md5 hash value of the keystore
+ * @property {string} [sha1] the sha1 hash value of the keystore
+ * @property {string} [sha256] the sha256 hash value of the keystore
+ * @property {string} [sha512] the sha512 hash value of the keystore
  */
 
 /**
  * Retrieve the the hash of the given keystore.
  *
- * @return {KeystoreHash}
+ * @return {Promise<KeystoreHash>}
  * @throws {Error} If getting keystore hash fails.
  */
 apkSigningMethods.getKeystoreHash = async function getKeystoreHash () {
@@ -359,3 +359,7 @@ apkSigningMethods.getKeystoreHash = async function getKeystoreHash () {
 };
 
 export default apkSigningMethods;
+
+/**
+ * @typedef {typeof apkSigningMethods} ApkSigningCommands
+ */
