+++ lib/tools/apks-utils.js
@@ -48,7 +48,7 @@ process.on('exit', () => {
  * @param {string} apks - The full path to the .apks file
  * @param {string|Array<String>} dstPath - The relative path to the destination file,
  * which is going to be extracted, where each path component is an array item
- * @returns {string} Full path to the extracted file
+ * @returns {Promise<string>} Full path to the extracted file
  * @throws {Error} If the requested item does not exist in the extracted archive or the provides
  * apks file is not a valid bundle
  */
@@ -90,7 +90,7 @@ async function extractFromApks (apks, dstPath) {
  *
  * @param {Array<String>} args - the list of bundletool arguments
  * @param {string} errorMsg - The customized error message string
- * @returns {string} the actual command stdout
+ * @returns {Promise<string>} the actual command stdout
  * @throws {Error} If bundletool jar does not exist in PATH or there was an error while
  * executing it
  */
@@ -129,7 +129,7 @@ apksUtilsMethods.execBundletool = async function execBundletool (args, errorMsg)
 
 /**
  * @param {string} specLocation - The full path to the generated device spec location
- * @returns {string} The same `specLocation` value
+ * @returns {Promise<string>} The same `specLocation` value
  * @throws {Error} If it is not possible to retrieve the spec for the current device
  */
 apksUtilsMethods.getDeviceSpec = async function getDeviceSpec (specLocation) {
@@ -146,18 +146,18 @@ apksUtilsMethods.getDeviceSpec = async function getDeviceSpec (specLocation) {
 
 /**
  * @typedef {Object} InstallMultipleApksOptions
- * @property {?number|string} timeout [20000] - The number of milliseconds to wait until
+ * @property {?number|string} [timeout=20000] - The number of milliseconds to wait until
  * the installation is completed
- * @property {string} timeoutCapName [androidInstallTimeout] - The timeout option name
+ * @property {string} [timeoutCapName=androidInstallTimeout] - The timeout option name
  * users can increase the timeout.
- * @property {boolean} allowTestPackages [false] - Set to true in order to allow test
+ * @property {boolean} [allowTestPackages=false] - Set to true in order to allow test
  * packages installation.
- * @property {boolean} useSdcard [false] - Set to true to install the app on sdcard
+ * @property {boolean} [useSdcard=false] - Set to true to install the app on sdcard
  * instead of the device memory.
- * @property {boolean} grantPermissions [false] - Set to true in order to grant all the
+ * @property {boolean} [grantPermissions=false] - Set to true in order to grant all the
  * permissions requested in the application's manifest automatically after the installation
  * is completed under Android 6+.
- * @property {boolean} partialInstall [false] - Install apks partially. It is used for 'install-multiple'.
+ * @property {boolean} [partialInstall=false] - Install apks partially. It is used for 'install-multiple'.
  * https://android.stackexchange.com/questions/111064/what-is-a-partial-application-install-via-adb
  */
 
@@ -165,7 +165,7 @@ apksUtilsMethods.getDeviceSpec = async function getDeviceSpec (specLocation) {
  * Installs the given apks into the device under test
  *
  * @param {Array<string>} apkPathsToInstall - The full paths to install apks
- * @param {?installMultipleApksOptions} options - Installation options
+ * @param {?InstallMultipleApksOptions} options - Installation options
  */
 apksUtilsMethods.installMultipleApks = async function installMultipleApks (apkPathsToInstall, options = {}) {
   const installArgs = buildInstallArgs(await this.getApiLevel(), options);
@@ -229,7 +229,7 @@ apksUtilsMethods.installApks = async function installApks (apks, options = {}) {
  * Extracts and returns the full path to the master .apk file inside the bundle.
  *
  * @param {string} apks - The full path to the .apks file
- * @returns {string} The full path to the master bundle .apk
+ * @returns {Promise<string>} The full path to the master bundle .apk
  * @throws {Error} If there was an error while extracting/finding the file
  */
 apksUtilsMethods.extractBaseApk = async function extractBaseApk (apks) {
@@ -243,7 +243,7 @@ apksUtilsMethods.extractBaseApk = async function extractBaseApk (apks) {
  * @param {string} apks - The full path to the .apks file
  * @param {?string} language - The language abbreviation. The default language is
  * going to be selected if it is not set.
- * @returns {string} The full path to the corresponding language .apk or the master .apk
+ * @returns {Promise<string>} The full path to the corresponding language .apk or the master .apk
  * if language split is not enabled for the bundle.
  * @throws {Error} If there was an error while extracting/finding the file
  */
@@ -276,3 +276,8 @@ apksUtilsMethods.isTestPackageOnlyError = function isTestPackageOnlyError (outpu
 };
 
 export default apksUtilsMethods;
+
+
+/**
+ * @typedef {typeof apksUtilsMethods} ApksUtils
+ */
