+++ lib/tools/settings-client-commands.js
@@ -624,7 +624,7 @@ commands.getSmsList = async function getSmsList (opts = {}) {
  * to the destination input field before this method is called.
  *
  * @param {string} text The string to type
- * @returns {boolean} `true` if the input text has been successfully sent to adb
+ * @returns {Promise<boolean>} `true` if the input text has been successfully sent to adb
  */
 commands.typeUnicode = async function typeUnicode (text) {
   if (_.isNil(text)) {
