+++ package.json
@@ -105,7 +105,7 @@
     "sinon": "^15.0.0",
     "temp": "^0.x",
     "ts-node": "^10.9.1",
-    "typescript": "^5.1.3"
+    "typescript": "~5.0.4"
   },
   "types": "./build/index.d.ts"
 }
